refactor: nest github mock under subfolder in E2E
