+++ e2e/github-device-flow/mock.ts
